requirements.txt
mlx
mlx-whisper
numpy
torch
torchaudio
openai
fastapi
uvicorn[standard]
websockets
requests
python-dotenv